+++ tests/testthat/test-main.R
@@ -1,6 +1,4 @@
 library(ggplot2)
-library(ggnormalviolin)
-library(testthat)
 d <- data.frame(
   Distribution = c("A", "B"),
   Distribution_mean = c(80, 90),
@@ -8,7 +6,8 @@ d <- data.frame(
 )
 
 test_that("example", {
-  expect_silent(ggplot(data = d, aes(x = Distribution)) +
-                  geom_normalviolin(aes(mu = Distribution_mean,
-                                        sigma = Distribution_sd)))
+  expect_silent(
+    ggplot(data = d, aes(x = Distribution)) +
+      geom_normalviolin(aes(mu = Distribution_mean, sigma = Distribution_sd))
+  )
 })